Fix styles 파일명 변경 -#12
